requirements.txt
# Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==5.0.0
Werkzeug==3.0.6
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43

# Authentication & Security
PyJWT==2.9.0
pyotp==2.9.0
qrcode==7.4.2

# API Documentation
flasgger==0.9.7.1

# Email Service
sendgrid==6.12.5

# Image Upload
cloudinary==1.44.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Security & Rate Limiting
Flask-Limiter==3.5.1

# Testing
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==5.0.0
coverage==7.4.2

# Optional development tools
ipython==8.12.0
